Added players List
previously Game Instance used its own Player object as a placeholder. Now the actual player object is implemented so a new List is added. It is now needed to change the placeholder Player Object with the actual implementation. To do so for now both lists are stored, but the placeholder will be replaced.