Inserted method to check if value can be insertd within floor
This is needed to determine whether a player is actually able to do a move with the new card combination. If he is able, he can do his round, if not a system error will be added. Depending on the amount of system errors it can also happen that a player has lost the game in this phase.