Removed Two Bugs of cheat Functionality
First Bug was a small fix, as the cheat button was enabled at every player field, even if the player is not present. This lead to a NullPointer Exception. Fixed this by checking whether the player is actually present or it is just a placeholder.

Second Bug was a misconception in the way of updating users about a successfull cheat detection or a unsuccessfull. Did this first by sending the hasCheated boolean via the success parameter, bought the implementation actually used this to only handle successfull requests. Now it is implemented by two different calls. One for the success case and one for the failure.